main.h: New header for extern variables and functions from main.cpp.
Moved all main.cpp externs from the various .cpp files into main.h
and consolidated some things.
